pyproject.toml
[project]
name = "upki-ca"
version = "0.1.0"
description = "uPKI CA instance"
authors = [
    {name = "x42en",email = "x42en@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "cryptography (>=46.0.5,<47.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "zmq (>=0.0.0,<0.0.1)"
]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0,<9.0.0"
pytest-cov = ">=6.0.0,<7.0.0"
pytest-asyncio = ">=0.25.0,<1.0.0"
ruff = ">=0.10.0,<1.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = []


[tool.poetry]
package-mode = true
packages = [{include = "upki_ca"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
